feat: added usage of token validation

+++ src/Dashboard/Profile/Experience/Education/FetchEducation/DeleteEducation.js
@@ -1,11 +1,12 @@
 const DeleteEducation = async (email)=>{
     const token = localStorage.getItem("token")
-    return await fetch(`http://localhost:8080/api/user-education/delete/${email}`, {
-        method: "DELETE",
-        headers:{
-            'Authorization' : `Bearer ${token}`
-        }
-    });
-
+    if(token) {
+        return await fetch(`http://localhost:8080/api/user-education/delete/${email}`, {
+            method: "DELETE",
+            headers: {
+                'Authorization': `Bearer ${token}`
+            }
+        });
+    }
 }
 export default DeleteEducation;
